+++ cli/list/mod.ts
@@ -3,13 +3,26 @@ import { getSupabaseClient } from '../utils.ts'
 import { login } from '../login/mod.ts'
 import { SUPABASE_FUNCTIONS_URL } from '../constants.ts'
 import { Arkive, Deployment } from '../../src/arkiver/types.ts'
+import { formatDistanceToNow } from 'npm:date-fns'
 
-export const action = async () => {
+type RawArkive = Omit<Arkive, 'deployment'> & {
+  deployments: (Omit<Deployment, 'created_at' | 'id'> & {
+    deployment_created_at: string
+    deployment_id: number
+  })[]
+  // deno-lint-ignore ban-types
+  environment: 'staging' | 'prod' | string & {}
+}
+
+export const action = async (options: {
+  all?: true
+  status?: string
+}) => {
   const dev = Deno.env.get('DEV') !== undefined
 
   if (dev) return listDev()
 
-  const spinner = wait('Fetching your arkives...').start()
+  let spinner = wait('Fetching your arkives...').start()
 
   try {
     // delete package
@@ -17,7 +30,9 @@ export const action = async () => {
     const sessionRes = await supabase.auth.getSession()
 
     if (!sessionRes.data.session) {
+      spinner.info('Not logged in, logging in now...')
       await login({}, supabase)
+      spinner = wait('Fetching your arkives...').start()
     }
 
     const userRes = await supabase.auth.getUser()
@@ -51,28 +66,62 @@ export const action = async () => {
 
     spinner.stop()
 
-    const arkives = (await listRes.json() as (Omit<Arkive, 'deployment'> & {
-      deployments: Deployment[]
-    })[]).flatMap((arkive) =>
-      arkive.deployments.map((deployment) => ({
-        name: arkive.name,
-        created_at: deployment.created_at,
-        id: arkive.id,
-        version: `${deployment.major_version}.${deployment.minor_version}`,
-        status: deployment.status,
-        is_public: arkive.public,
-      }))
-    )
+    const rawArkives = await listRes.json()
 
-    console.table(arkives)
+    if (options.all) {
+      const arkives = (rawArkives as RawArkive[]).flatMap((arkive) =>
+        arkive.deployments.map((deployment) => ({
+          name: arkive.name,
+          deployed: `${
+            formatDistanceToNow(
+              new Date(deployment.deployment_created_at),
+            )
+          } ago`,
+          version: `${deployment.major_version}.${deployment.minor_version}`,
+          status: deployment.status,
+          arkive_id: arkive.id,
+          deployment_id: deployment.deployment_id.toString(),
+        })).filter((deployment) =>
+          options.status ? deployment.status === options.status : true
+        )
+      )
 
-    Deno.exit()
+      console.log('All deployments:')
+      console.table(arkives)
+    } else {
+      const arkives = (rawArkives as RawArkive[]).map((arkive) => {
+        const latestDeployment = arkive.deployments.sort((a, b) =>
+          a.deployment_id - b.deployment_id
+        )[0]
+
+        return {
+          name: arkive.name,
+          deployed: `${
+            formatDistanceToNow(
+              new Date(latestDeployment.deployment_created_at),
+            )
+          } ago`,
+          version:
+            `${latestDeployment.major_version}.${latestDeployment.minor_version}`,
+          status: latestDeployment.status,
+          endpoint: craftEndpoint({
+            arkiveName: arkive.name,
+            environment: arkive.environment,
+            majorVersion: latestDeployment.major_version,
+            username,
+          }),
+        }
+      }).filter((deployment) =>
+        options.status ? deployment.status === options.status : true
+      )
+
+      console.log('Latest deployments:')
+      console.table(arkives)
+    }
   } catch (error) {
-    spinner.fail('Deletion failed: ' + error.message)
-    console.error(error)
+    spinner.fail('Listing arkives failed: ' + error.message)
+    return
   }
-
-  Deno.exit()
 }
 
 const listDev = async () => {
@@ -117,4 +166,22 @@ export const getUsername = async (userId: string) => {
   }
 
   return profileRes.data.username
+}
+
+export const craftEndpoint = (
+  params: {
+    username: string
+    arkiveName: string
+    // deno-lint-ignore ban-types
+    environment: 'prod' | 'staging' | string & {}
+    majorVersion: number
+  },
+) => {
+  const { arkiveName, environment, username, majorVersion } = params
+
+  const baseGraphQlUrl = environment === 'prod'
+    ? `https://data.arkiver.net/${username}`
+    : `https://data.staging.arkiver.net/${username}`
+
+  return `${baseGraphQlUrl}/${arkiveName}/${majorVersion}/graphql`
 }